+++ http4k-security-oauth/src/main/kotlin/org/http4k/security/oauth/server/request/RequestObjectExtractor.kt
@@ -1,6 +1,14 @@
 package org.http4k.security.oauth.server.request
 
+import com.fasterxml.jackson.annotation.JsonInclude
+import com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL
 import com.fasterxml.jackson.core.JsonParseException
+import com.fasterxml.jackson.databind.DeserializationFeature
+import com.fasterxml.jackson.databind.DeserializationFeature.FAIL_ON_IGNORED_PROPERTIES
+import com.fasterxml.jackson.databind.DeserializationFeature.FAIL_ON_MISSING_EXTERNAL_TYPE_ID_PROPERTY
+import com.fasterxml.jackson.databind.DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES
+import com.fasterxml.jackson.databind.DeserializationFeature.USE_BIG_DECIMAL_FOR_FLOATS
+import com.fasterxml.jackson.databind.DeserializationFeature.USE_BIG_INTEGER_FOR_INTS
 import com.fasterxml.jackson.module.kotlin.KotlinModule
 import com.natpryce.Failure
 import com.natpryce.Result
@@ -53,6 +61,13 @@ object RequestObjectExtractor {
             .text(BiDiMapping(State::class.java, { State(it) }, { it.value }))
             .text(BiDiMapping(Nonce::class.java, { Nonce(it) }, { it.value }))
             .done()
+            .setSerializationInclusion(NON_NULL)
+            .configure(FAIL_ON_UNKNOWN_PROPERTIES, false)
+            .configure(FAIL_ON_IGNORED_PROPERTIES, false)
+            .configure(USE_BIG_DECIMAL_FOR_FLOATS, true)
+            .configure(USE_BIG_INTEGER_FOR_INTS, true)
+            .configure(FAIL_ON_MISSING_EXTERNAL_TYPE_ID_PROPERTY, false)
+
     )
 
 }